+++ test/xml/xml_write_read_test.cc
@@ -55,6 +55,7 @@ std::vector<std::string> GetWriteReadTestModels() {
             absl::StrContains(xml, "rfcamera") ||
             absl::StrContains(xml, "tactile") ||
             absl::StrContains(xml, "makemesh") ||
+            absl::StrContains(xml, "carousel") ||
             absl::StrContains(xml, "many_dependencies") ||
             absl::StrContains(xml, "usd") ||
             absl::StrContains(xml, "torus_maxhull") ||